Use seaparate component for attachment cards
