+++ components/Attachments.js
@@ -1,36 +1,25 @@
 import React from "react";
-import { View, StyleSheet } from "react-native";
+import { View } from "react-native";
 
-import { Card, Text, IconButton, useTheme } from "react-native-paper";
+import { Chip } from "react-native-paper";
 
 import { openFile } from "../util/files.js";
 import api from "../util/api";
 
 export default function Attachments({ files }) {
-  const theme = useTheme();
-
-  const styles = StyleSheet.create({
-    fileName: {
-      color: "white",
-      alignSelf: "center",
-      flex: 1,
-    },
-    fileIcon: {
-      margin: 0,
-      alignSelf: "center",
-    },
-    card: {
-      padding: 10,
-      backgroundColor: theme.colors.primary,
-      marginBottom: 8,
-    },
-  });
-
   return (
-    <View>
+    <View
+      style={[
+        {
+          flexDirection: "row",
+          flexWrap: "wrap",
+        },
+      ]}
+    >
       {files.map((file, index) => (
-        <Card
+        <Chip
           key={index}
+          icon="file-outline"
           mode="outlined"
           onPress={() => {
             openFile(
@@ -39,17 +28,10 @@ export default function Attachments({ files }) {
               file.name
             );
           }}
-          style={styles.card}
+          style={{ margin: 4, width: "97%" }}
         >
-          <View style={{ flexDirection: "row" }}>
-            <IconButton
-              icon="file"
-              color="white"
-              style={styles.fileIcon}
-            ></IconButton>
-            <Text style={styles.fileName}>{file.name}</Text>
-          </View>
-        </Card>
+          {file.name}
+        </Chip>
       ))}
     </View>
   );
